requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==2.2.17
django-cors-headers==3.7.0
django-filter==2.4.0
django-registration-redux==2.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
Pastebin==1.1.0
Pillow==8.1.0
pylint==2.5.3
pytz==2020.5
requests==2.25.1
rope==0.18.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.26.3
wincertstore==0.2
wrapt==1.12.1
